refactor: merge repos and services
