+++ crates/length-log-core/src/service.rs
@@ -1,66 +1,24 @@
 use std::{fmt, str::FromStr};
 
-use chrono::{Local, NaiveDate};
-// use crate::error::Error;
-// use nom::{bytes::complete::tag, combinator::map, IResult};
-// use polars::prelude::DataFrame;
-
-// fn person(i: &str) -> IResult<&str, &str> {
-//     tag("person")(i)
-// }
-
-// pub fn handle_command(command: String) -> Result<String, Error> {
-//     log::trace!("handle command: {}", command);
-//     let cmd = parse_line(&command)?;
-//     match cmd {
-//         Command::Person => {
-//             log::debug!("got {:?}", cmd);
-//         }
-//     }
-//     todo!("handle_command")
-// }
-
-// pub fn parse_line(line: &str) -> Result<Command, Error> {
-//     log::trace!("parse line: {}", line);
-//     let res = map(person, |_: &str| Command::Person)(line);
-//     match res {
-//         Ok((_, cmd)) => Ok(cmd),
-//         Err(err) => Err(Error::Unknown(format!("{}", err))),
-//     }
-// }
-
-// #[derive(Debug, Clone)]
-// pub enum Command {
-//     Person,
-// }
-
 use crate::{
     models::{self, AddPersonRequest, Person},
-    ports::{
-        data::{DataRepository, DataService},
-        person::{PersonRepository, PersonService},
-        LengthLogService, ServiceError,
-    },
+    ports::{LengthLogRepository, LengthLogService, ServiceError},
 };
 
 #[derive(Clone)]
-pub struct Service<PR, DR> {
-    person_repo: PR,
-    data_repo: DR,
+pub struct Service<R> {
+    repo: R,
 }
 
-impl<PR, DR> Service<PR, DR> {
-    pub fn new(person_repo: PR, data_repo: DR) -> Self {
+impl<R> Service<R> {
+    pub fn new(repo: R) -> Self {
         log::trace!("creating Service ...");
-        Self {
-            person_repo,
-            data_repo,
-        }
+        Self { repo }
     }
 }
-impl<PR, DR> PersonService for Service<PR, DR>
+impl<R> LengthLogService for Service<R>
 where
-    PR: PersonRepository,
+    R: LengthLogRepository,
 {
     fn add_person(&self, req: models::AddPersonRequest) -> Result<Person, models::AddPersonError> {
         let person: Person = req.into();
@@ -70,71 +28,25 @@ where
             person.start_date()
         );
 
-        self.person_repo.save(&person)?;
+        self.repo.save_person(&person)?;
         Ok(person)
     }
-}
 
-impl<PR, DR> DataService for Service<PR, DR>
-where
-    DR: DataRepository,
-{
     fn add_datapoint(
         &self,
         req: models::datapoint::AddDatapointRequest,
     ) -> Result<(), ServiceError> {
         todo!()
     }
-}
 
-impl<PR, DR> LengthLogService for Service<PR, DR>
-where
-    PR: PersonRepository,
-    DR: DataRepository,
-{
     fn list_persons(&self) -> Result<Vec<Person>, ServiceError> {
         todo!()
     }
     fn save(&self) -> miette::Result<()> {
-        self.person_repo.dump()?;
-        self.data_repo.dump()?;
+        self.repo.dump()?;
         Ok(())
     }
 }
-//     pub fn add_person(&self, name: String, start_date: Option<String>) -> Result<(), AppError> {
-//         log::trace!("adding person '{}' with date = {:?}", name, start_date);
-//
-//         let start_date = if let Some(start_date_str) = start_date {
-//             Some(NaiveDate::from_str(&start_date_str)?)
-//         } else {
-//             None
-//         };
-//         let person = Person::with_name_and_start_date(name, start_date);
-//         self.person_service.save(person)?;
-//         Ok(())
-//     }
-//
-//     pub fn list_persons(&self) -> Result<Vec<models::Person>, AppError> {
-//         Ok(self.person_service.get_all()?)
-//     }
-//
-//     pub fn add_data(&self, name: &str, date: Option<String>, data: f64) -> Result<(), AppError> {
-//         log::trace!(
-//             "adding datapoint for person '{}' with date = {:?}",
-//             name,
-//             date
-//         );
-//         let id = self.person_service.get_id_by_name(name)?;
-//         dbg!(&id);
-//         let date = if let Some(date_str) = date {
-//             NaiveDate::from_str(&date_str)?
-//         } else {
-//             Local::now().naive_local().date()
-//         };
-//         self.data_service.save(&id, date, data)?;
-//         Ok(())
-//     }
-// }
 
 #[derive(Debug)]
 pub enum AppError {
