+++ crates/length-log/src/main.rs
@@ -2,7 +2,7 @@ use std::fs;
 
 use directories::ProjectDirs;
 use length_log_core::Service;
-use length_log_polars::{datapoints::PolarsDataRepo, persons::PolarsPersonRepo};
+use length_log_polars::repo::PolarsRepository;
 use length_log_repl::{ReplApp, ReplAppConfig};
 use miette::IntoDiagnostic;
 
@@ -20,17 +20,15 @@ fn main() -> miette::Result<()> {
     let data_path = data_dir.join("data.parquet");
     let history_path = data_dir.join("history");
 
-    let person_repo = if persons_path.exists() {
-        PolarsPersonRepo::from_path(persons_path.as_path()).into_diagnostic()?
+    let repo = if persons_path.exists() && data_path.exists() {
+        PolarsRepository::from_path(data_path, persons_path).into_diagnostic()?
     } else {
-        PolarsPersonRepo::with_path(persons_path)
+        if persons_path.exists() || data_path.exists() {
+            todo!();
+        }
+        PolarsRepository::with_path(data_path, persons_path)
     };
-    let data_repo = if data_path.exists() {
-        PolarsDataRepo::from_path(data_path).into_diagnostic()?
-    } else {
-        PolarsDataRepo::with_path(data_path)
-    };
-    let service = Service::new(person_repo, data_repo);
+    let service = Service::new(repo);
     let app_config = ReplAppConfig { history_path };
     let mut app = ReplApp::new(service, app_config)?;
     app.run()?;
